Fix the getTime() function again.

+++ src/core/engine.c
@@ -40,7 +40,7 @@ static uint64_t getTime()
 {
 	struct timeval tv;
 	gettimeofday(&tv, 0);
-	return tv.tv_sec * 1000000 + tv.tv_usec;
+	return (uint64_t) tv.tv_sec * 1000000 + tv.tv_usec;
 }
 
 void yukonEngineCapture(struct yukonEngine *engine)
